Batang ready to go!
This one is looking good too! Again, made a slight modification to the ogonek character.

Also, I merged the Regular and Bold together into a single Glyphs file—personal preference, but I find that this helps ensure that the two files are kept properly in sync with one another in terms of font metadata and other bits.